requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
cloudinary==1.41.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fsspec==2024.10.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pillow==11.0.0
PyJWT==2.10.0
python-dotenv==1.0.1
setuptools==75.5.0
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.1
torch==2.5.1
torchgen==0.0.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
